+++ crates/node/src/service/mod.rs
@@ -1,7 +1,9 @@
 mod block_production;
+mod l1;
 mod rpc;
 mod sync;
 
 pub use block_production::BlockProductionService;
+pub use l1::L1SyncService;
 pub use rpc::RpcService;
 pub use sync::SyncService;